requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.4.6
python-dotenv==1.0.1
jinja2==3.1.3
pydantic-settings==2.2.1
